+++ lib/serega/plugins/validate_modifiers/validate.rb
@@ -1,6 +1,9 @@
 # frozen_string_literal: true
 
 class Serega
+  class AttributeNotExist < Error
+  end
+
   module Plugins
     module ValidateModifiers
       class Validate
@@ -29,14 +32,14 @@ class Serega
           def raise_error(name, prev_names)
             field_name = field_name(name, prev_names)
 
-            raise Serega::Error, "Attribute #{field_name} not exists"
+            raise Serega::AttributeNotExist, "Attribute #{field_name} not exists"
           end
 
           def raise_nested_error(name, prev_names, nested_fields)
             field_name = field_name(name, prev_names)
             first_nested = nested_fields.keys.first
 
-            raise Serega::Error, "Attribute #{field_name} is not a relation to add '#{first_nested}' attribute"
+            raise Serega::AttributeNotExist, "Attribute #{field_name} has no :serializer option specified to add nested '#{first_nested}' attribute"
           end
 
           def field_name(name, prev_names)