+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Plugin `validate_modifiers` now raises `Serega:AttributeNotExist` error when requested attribute not exists
+
 - Change `preloads` plugin config option `config[:preloads][:auto_preload_relation] = true` to `config[:preloads][:auto_preload_attributes_with_serializer] = false`, so now there are no surprises where this preloads come from.
 
 ## [0.1.4] - 2022-08-25